Combine/Hackathon/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=1.4
psycopg2-binary>=2.9
python-dotenv>=1.0
passlib[bcrypt]>=1.7
python-jose[cryptography]>=3.3
pydantic>=1.10,<2.0
alembic>=1.11
sentence-transformers
scikit-learn
numpy
torch
